Move src/loading to src/core/loading-instructions
